+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add `salmon.util/pages-seq` for handling paginated responses.
 - Add `salmon.cleanup/deregister-all-amis!` for cleaning up AMIs in test accounts.
 - Add `salmon.cleanup/delete-orphaned-snapshots!` for cleaning up snapshots in test accounts.
+- Use `:cognitect.aws.error/code` to find the error code in `salmon.util/aws-error-code`.
 
 ## v0.19.0 (2024-09-16)